requirements.txt
mysqlclient
djangorestframework
django-filter
